event_manager/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile '.\requirements.in'
#
asgiref==3.8.1
    # via django
crispy-bootstrap5==2024.10
    # via -r .\requirements.in
django==5.1.3
    # via
    #   -r .\requirements.in
    #   crispy-bootstrap5
    #   django-crispy-forms
django-crispy-forms==2.3
    # via
    #   -r .\requirements.in
    #   crispy-bootstrap5
django-environ==0.11.2
    # via -r .\requirements.in
factory-boy==3.3.1
    # via -r .\requirements.in
faker==33.1.0
    # via factory-boy
python-dateutil==2.9.0.post0
    # via faker
six==1.16.0
    # via python-dateutil
sqlparse==0.5.2
    # via django
typing-extensions==4.12.2
    # via faker
tzdata==2024.2
    # via django
whitenoise==6.8.2
    # via -r .\requirements.in
